refactor: common 폴더 파일 이동
- dto, filter => presentation 폴더로 묶기
- exception, factory => infra 폴더로 묶기
