+++ src/coupon/infrastructure/persistence/factories/coupon.factory.ts
@@ -2,7 +2,7 @@ import { CouponTypeOrmEntity } from "../orm/coupon.typeorm.entity";
 import {
   createEntityFactory,
   getBaseProps,
-} from "../../../../common/factories/base.factory";
+} from "../../../../common/infrastructure/base.factory";
 import {
   TEST_FACTORY_DEFAULTS,
   createTestName,
