+++ src/coupon/infrastructure/persistence/factories/user-coupon.factory.ts
@@ -2,7 +2,7 @@ import { UserCouponTypeOrmEntity } from "../orm/user-coupon.typeorm.entity";
 import {
   createEntityFactory,
   getBaseProps,
-} from "../../../../common/factories/base.factory";
+} from "../../../../common/infrastructure/base.factory";
 import { TEST_FACTORY_DEFAULTS } from "./constants";
 import { v4 as uuidv4 } from "uuid";
