+++ src/order/presentation/http/filters/order-exception.filter.ts
@@ -10,7 +10,7 @@ import { ErrorCode } from "@/common/constants/error-codes.enum";
 import {
   BaseExceptionFilter,
   ErrorMapping,
-} from "@/common/filters/base-exception.filter";
+} from "@/common/presentation/filters/base-exception.filter";
 
 /**
  * 주문 도메인 예외 처리 필터
