+++ src/product/presentation/http/filters/product-exception.filter.ts
@@ -11,7 +11,7 @@ import { ErrorCode } from "@/common/constants/error-codes.enum";
 import {
   BaseExceptionFilter,
   ErrorMapping,
-} from "@/common/filters/base-exception.filter";
+} from "@/common/presentation/filters/base-exception.filter";
 import {
   ProductNotFoundError,
   StockReservationNotFoundError,
