+++ src/wallet/presentation/http/filters/wallet-exception.filter.ts
@@ -9,7 +9,7 @@ import { ErrorCode } from "@/common/constants/error-codes.enum";
 import {
   BaseExceptionFilter,
   ErrorMapping,
-} from "@/common/filters/base-exception.filter";
+} from "@/common/presentation/filters/base-exception.filter";
 import {
   UserBalanceNotFoundError,
   PointTransactionNotFoundError,
