+++ src/wallet/presentation/http/wallet.controller.ts
@@ -18,7 +18,7 @@ import {
   BalanceResponseDto,
   ChargeResponseDto,
 } from "./dto/wallet.dto";
-import { ApiResponseDto } from "../../../common/dto/response.dto";
+import { ApiResponseDto } from "../../../common/presentation/dto/response.dto";
 import { JwtAuthGuard } from "@/auth/guards/jwt-auth.guard";
 import {
   CurrentUser,
